// plugins/repositories-plugin/src/types.ts
export interface GitHubRepository {
  name: string;
  full_name: string;
  description: string;
  html_url: string;
  stargazers_count: number;
  forks_count: number;
  language: string;
  owner: {
    login: string;
    avatar_url: string;
  };
}

export interface RepositoriesData {
  repositories: GitHubRepository[];
}